requirements.txt
pandas
numpy
matplotlib
pyyaml
reportlab
schedule
python-dotenv
